Major Cleanup and the Introduction of the Relations Constructor Function
Renamed all _id properties in the various API constructors to _class to ensure no conflicts for a later change regarding Hypermedia Control IDs.
Renamed Resources.list to Resources.getAffordances to ensure no confusion when we later introduce option flags to the list command.
Renamed a ton of references to file path variables to make for easier readability.
Removed the contents of the unit test command. The tests were outdated, and should be updated later anyways.
Introduced the Relations Constructor Function to allow for checking if relations entered in the CLI exist against the IANA specification. A module-scoped reference to an instance of this Constructor was also introduced.
Made the path string cleansing for the response file paths much more readable.
Introduced an -i and --id options flag. This is needed for further changes being made regarding the Hypermedia Control IDs.