Introduced the notion of the ID to the Method constructor.
ID is completely optional. If the argument is passed in as a string, then it is used. Else, it defaults the property to undefined. This is done to prevent serialization of empty string ID fields. IDs should be explicitly, globally unique.
The extra ID argument is introduced in the convenience, instantiator method exposed in the exports object.