Introduced various methods to the Resources prototype object to allow for management of Affordances by ID.
Resources.getMethod allows for the retrieval of a Method on any Resource in the API from the root level by ID.
Resources.deleteMethod allows for the deletion of a Method on any Resource in the API from the root level by ID.
Resources.getIds returns an Array of all existing Method IDs already defined in the API specification.

get/deleteMethod were necessary to allow for command-line invocation of get/del without the need to specify the resource as an argument.The --id option should be thought of as analogous to stating the resource and method explicitly
because resource/method pairs are implicitly globally unique within the API specification and method IDs are explicitly globally unique within the API specification.

Also, started passing in the Method IDs when instantiating an Affordance in Resources.getAffordances.