Added the ability to ID your affordances in your API specification prior to attempting generation of responses.
The -i and --id options were introduced.
This option is allowed when put, get, del, and gen are invoked.
In the context of put, --id is just an ID for the Method instance being put on any predefined Resource.
In the context of get, --id is used in place of typing [res] [met] because it is semantically the same thing.
In the context of del, --id is used in place of typing [res] [met] because it is semantically the same thing.
In the context of gen, --id is used in place of typing [res] [met] because it is semantically the same thing.

Examples:
put / GET --id entry
get --id entry
del --id entry
gen --id entry --returns 200

Modifications were also made the the Generator delegate callback to only prompt for an ID if one was NOT specified in the API specification document, and to removed previously chosen action possibilities at runtime.
Various cleanup was also done.