Removed all comments for consolidation purposes during development because the file was getting bloated. Will reintroduce them later.
Made all paths absolute. Relative paths are allowed with using the accessor or constructor, but they will be made absolute under the hood for disambiguation's sake.

Introduced the notion of a captain's log. This allows me to log every encounter of a file, symbolic link, directory, character device, block device, pipe, or socket into their respective Arrays.
These Arrays are referenced on each instance of a Walker, Texas Ranger as a field on the object.
This logging may be turned on and off with an accessor setShouldLogWalk(should), and may be introspected with getShouldLogWalk().

Modified the unit testing. These should be removed relatively soon. I'll remove them when I'm about to incorporate this module into the prototype of the server.