Removed the module reference to the file name string in favor of inlining the string.
Renamed the 'api' folder to 'gen' to keep the naming convention more inline with other development environments.

Removed the module reference to affordances in favor of invoking api.getAffordances when it's needed. This keeps the array up-to-date instead of maintaining a stale reference.

Modified the ls command to remind me to add flags for various types of lists.

Started writing a second file in the 'gen' folder for exposing all action possibilities currently available in the API. The file is called actions.json.