modified:   lib/affordance.js modified:   lib/affordances.js modified:   lib/input.js
Added canRevive and revive methods to the exports object in each respective module.
CanRevive takes an arbitrary JS Object, and duck-type checks it to see if it meets the requirements necessary to be revived as an instance.
Revive takes the arbitrary JS Object, and creates an instance of the respective module's constructor and copies the fields of the JS Object over. This acts like a copy constructor.