modified:   lib/input.js
Accept and Required are no longer required arguments in an Input constructor. They can still be added with accessor methods.
Id and Value are still required arguments, however.

The removed arguments are reflected in the Constructor signature, the convenience wrapper function signature, the canRevive method definition, and accompanying documentation.
setRequired's definition was also changed to only allow setting a value if it's of boolean type.