Modified node_modules/xhn/lib/affordances.js
Added a handful of examples explaining the motivation and use cases for metadata cascading within collections of Affordances. It's analogous to CSS, but for HTTP request headers.
Changed a considerable amount of the method documentation.
Added a method to the Affordances.prototype Object called cascadeMetadata.
The method allows Affordances collection metadata to merge with a contained child Affordance or Affordances instances metadata.
Added a method to the Affordances.prototype Object called copy.
This method allows you to make a shallow-copy of an Affordances instance in place.
We are now utilizing copy and cascadeMetadata in getAffordanceById. The documentation was changed to reflect this subtle change.
The reason we are no longer returning explicit references to Affordance(s) found by id is because it allows us to modify the instances without regard to future side-effects of successive searches for the same ids. If we make changes to copies, then it shouldn't impact anything.