modified:   common.js new file:   message.js
Modified the common module to extend interfaces for the Message, Messages, and Hapi modules in the exports object.
Added Message so far. Messages and Hapi will come soon.

Message represents an Affordance-Rich Message. This is effectively representing nodes on a HyperGraph. This is the fourth interface constraint of REST, Hypermedia as the Engine of Application State.
Messages will be a collection of nodes from a Hypermedia API (Hapi).