modified:   node_modules/xhn/lib/affordances.js
Removed the giant comment regarding Metadata Cascading Principles.
Moved the information to the README.md file and formatted the text in Markdown syntax.
Added multiple methods including: getAffordanceAtIndex, getLastAffordance, hasAffordanceWithId
getAffordanceAtIndex allows for the retrieval of Affordance(s) from a collection at a specified location.
getLastAffordance leverages getAffordanceAtIndex and just returns whatever it finds at the last possible index in the Array, length - 1.
hasAffordanceWithId introspects the collection of Affordance(s) for anything that may have a specified identifier, and it returns a boolean indicating if it found something that matched. This allows for global uniqueness checks within a collection.