modified:   node_modules/xhn/lib/message.js modified:   node_modules/xhn/lib/messages.js
Improved the argument checking in the Message constructor and accessor method. It now also checks that the Array that represents the message has at least one index.
It is now also possible to check if a Message is meant for a particular request/response exchange with the newly introduced isForExchange method.

It is now possible to enumerate all Messages in a Messages instance, and decide whether any contained Messages are for a particular request/response exchange with the newly introduced hasMessageForExchange method.