deleted:    hypermediator.js
Removing Hypermediator in favor of having instances of HypermediaApi do the 'Hypermediation.'
The code was slimmed down considerably because the argument checking that went into the mediate method prior is no longer necessary with an instance of HypermediaApi.
Since the introduction of the Message constructor took place, adding responses to HypermediaApis before attempting to create runtime responses takes care of the need to have argument validation.