new file:   hapi.js
Adding the HypermediaApi Constructor to the XHN library of Constructors.

HypermediaApi allows for the instrumentation of a Hypermedia API by wrapping the facilities provided by the other Constructors located within the library path.
In other words, it allows you to create an instance of a Hypermedia API that has a particular:Transfer Protocol, Hypermedia-Aware Media Type Encoding,
Name, Set of Request Possibilities, and Set of Response Possibilities.
It's meant to be treated as a Singleton instance whenever a live HTTP API server is running to allow for the API server to respond with Hypermedia responses.