modified:   xhb/lib/common.js
Added three modules as properties on the exports Object of the common module.
These include: Argo, Hypermediator, Plugin.

Hypermediator allows for the creation of an instance of Hypermediator. Hypermediators allow you to transform an instance of HypermediaApi to a live HTTP server.
Plugin allows for the creation of an instance of Plugin. These allow you to decide what framework to bridge the HypermediaApi instance to. This project is beginning development by utilizing Argo. ExpressJS may be added later.
Argo is a module that exports an instance of Plugin developed particularly for bridging a HypermediaApi to an instance of Argo.