modified:   xhb/package.json
Changed the name of the module to simply, 'xhb.'
This was done to conform to NPMs naming conventions for modules.

Also added argo-url-router as a dependency for the project.
Argo-url-router allows for friendlier URI paths to be used as opposed to Regular Expression Objects in the routing system of Argo.