modified:   ../xhb/lib/hypermediator.js modified:   ../xhb/lib/plugin.js
Modified Hypermediator.mediate to return true if an HTTP was spawned using a Plugin extension, and false otherwise. It used to return undefined (nothing).
Modified Plugin.bridge to return true if an HTTP server was spawned using the delegate function field within the instance, and false otherwise. It used to return undefined as well.
Any delegate callback function set in a Plugin instance should return true/false indicative of whether the delegate was capable of creating an HTTP server from an instance of HypermediaApi passed to the function invocation.