modified:   ../xhb/lib/plugin.js
Changed the return of Plugin.bridge to be 'return !!delegate(api);' just in case the developer of a Plugin fails to return a valid boolean.
If they fail to return anything, the undefined return value gets flagged to false because it is falsy.
If they return something other than a boolean, then it is changed to a boolean dependant on its truthiness.