modified:   ../xhn/lib/hapi.js
Added the 'canHost' method to the HypermediaApi prototype object.
This method allows for checking if the HypermediaApi meets the minimum constraints necessary for being hosted as a fully hypermedia-capable API server.
This is being used in the hypermediator to decide whether or not the Plugin extensions should even waste the effort in attempt to start a server from a HAPI DSL.