modified:   package.json
Removed the flag to treat it as a private module regarding NPM.
Bumped the node engine to 0.10.x. This is to make sure that the version of Node we use for the Argo dependency is correct.
Bumped the argo dependency version to 0.4.9 to make sure the PR we created was merged in.