modified:   lib/hypermediator.js modified:   lib/plugin.js
Hypermediator instances now accept a port and host string to assist in the creation of web server TCP/IP sockets.
The port and host are passed to the Plugin instance alongside the HypermediaApi in the Plugin.bridge method invocation during mediation.
Plugin now expects the extra arguments in its bridge method and passes them through to the bridge delegate function.