modified:   affordances.js
Added accessor methods to the Affordances Prototype to allow for retrieval of fields on an instance of Affordances including Id, and Metadata.
Added a forEach method variation to the Affordances Prototype to allow for delegate callback functions to be invoked with every encountered Affordance within a collection of Affordances. This assists primarily with mapping routes to an HTTP routing framework such as Argo or Express.