modified:   index.js
Modified XOP heavily.

Did some major documentation consolidation and cleanup.
Removed the concept of the options Array when instantiating an Options Object.The parser no longer cares about what options you may be acknowldging. It now just keeps track of any argument that looks like an option based on the RegExp style. Whether or not an option was meant to be introduced to the argv is up to the developer.