deleted:    message.js deleted:    messages.js
Removed the Message and Messages Modules because they are no longer relevant to the project.
The notion of Affordance-Rich Messages still exists. However, the messages themselves are now maintained within Affordance instances to heavily condense the logic and make access to the messages much easier within request handlers. It removes the level of indirection that was necessary through a HypermediaApi instance prior.