modified:   affordances.js modified:   input.js
Added toJSON methods to the Prototype Objects of Affordances and Input Constructors to make sure that all instances can be stringified with only what is necessary to describe the instance in NavAL+JSON media type semantics.
This is necessary for allowing us to add fields to the instances at runtime that are NOT intended to be in the NavAL+JSON media type serializations.