deleted:    sshd_config new file:   sshd_config.awk new file:   sshd_config.man
Removed the sshd_config file since it doesn't really make sense to modify the entire file from the default.
Wrote an AWK script to match key/value pairs in the default sshd_config file, and replace the values with my own default values.
Included a manpage file for sshd_config in the path.