modified:   Dockerfile
Changed the Dockerfile to make the SSHD configuration more flexible.
An AWK file modifies the sshd_config file and outputs a new sshd_config while retaining a backup of the old one that was installed with ssh-server.
The git remote is renamed to default.git from hapi.git to make the image more agnostic.
An sshd_config file is no longer maintained in the project repository. It was removed within two previous commits.
An sshd_config manpage file is now present in the project, but it is NOT added to the Docker images.