deleted:    image/git-shell-commands/addkey deleted:    image/git-shell-commands/create new file:   image/git-shell-commands/git modified:   image/git-shell-commands/help new file:   image/git-shell-commands/key deleted:    image/git-shell-commands/list
Removed, Modified, and Created git-shell-commands Bourne Shell scripts.
addkey was removed in favor of 'key --add'
create was removed in favor of 'git --add'
list was removed in favor of 'git --list'
git and key scripts were added to introduce those new commands. The scripts are simplified from the previous states as well.
help was simplified heavily as well.