Added Dockerfile to Hapi.
The Dockerfile builds an image based on 'hapi/hapi.'
The Dockerfile adds the entire project to the image at a directory /home/git/app, then the Git user creates a new, local Git repository, and pushes to /home/git/app.git as the remote.
This allows the source of Hapi to be immediately accessible by every Hapi container created.
Users cloning their own app.git from their remotes will be able to directly modify the source that runs a Hapi, and create pull requests for the official remote on GitHub.