modified: Makefile
Modified the Makefile to ensure logs were tailed to a logfile, forever doesn't start a failing application more than 5 times, and pass new options to the application.
The new options make the host more verbose to 0.0.0.0, and add a routes parameter to supply an explicit path for the request handlers of a Hapi.