deleted: node_modules/xop/
Removed XOP in favor of publishing it to npm under a new name, Slop.
It's the same source, but is now accessible from npm instead.
With this module removed, we are now able to completely ignore the nested node_modules path in favor of installing dependencies with facilitation by package.json.