#deleted:    api/api.js #new file:   api/api.json #new file:   api/directory/structure/doesnt/matter/entry.js #new file:   api/directory/structure/payment.js #new file:   api/useragent.js
Deleted api.js as we are no longer supporting Javascript as the internal Domain Specific Language. We are now parsing JSON instead.
Introducing api.json to replace api.js. It is an internal DSL to represent API descriptions (similar to API Blueprint, Swagger, RAML, etc).
Attempted to showcase most of the feature-set of api.json as default examples in the file.
Attempted to showcase a convention of disregarding the directory tree structure within the directory that contains api.json.
node index.js --dir=api/directory/ should be the root of any directory that contains an api.json file and accompanying request handler module.js files.
The directory will be traversed recursively until everything is found, so their respective locations are irrelevant with respect to the root directory.
entry.js, payment.js, and useragent.js are examples of request handler modules for the api.
The entry point handler just includes hypermedia semantics.
The payment handler includes application data and hypermedia semantics.
The useragent handler is a convention for returning the user agent that will traverse the hypermedia api directly from the hypermedia api origin itself.
This is purely a convention, and does not have to happen from all origins made by this software.