modified:   package.json
Modified the package.json file for various reasons.
Changed the project name from hapi to coast.
Changed the description because the purpose of the project has now changed dramatically.
Updated the license from GPL2 to MIT.
Bumped the version of the project from 0.0.1 -> 1.0.0 because breaking changes have been made.
Added a binary to be installed at {prefix}/bin for running coast from the CLI.
Updated the engine to >=0.10 as opposed to the 0.10.x restriction.
Alphabetized the dependency list.
Updated all URLs to the coast Github name.
Added keywords.