Command Line Interface Changes.
The interface is now accessible from your PATH as a node.js global executable.
The interface no longer accepts --debug and --dir. --debug has moved to a debug field in api.json. --dir is removed in favor of always using cwd.
Handlers no longer use the Argo format. They use Express now as more people are familiar with that style.
The interface now searches the cwd for a file called api.json and creates an Express server from the description.
The interface also searches for files in cwd for api and middleware handlers of the naming convention [id].js.
Slop was removed as a dependency because there is no longer a need for option flags, like --debug.
Argo built-in middleware was converted to Express middleware, like CORS, hypermedia, etc.
UglifyJS was added as a dependency because the hypermedia client is now built into a default resource on any coast server, /ua. The returned client is minified Javascript.
HTTP was removed as a dependency because Express .listen() is used which creates an HTTP server internally.